StroiNadzorAI-main/requirements.txt
# ===========================================
# ОБЯЗАТЕЛЬНЫЕ ЗАВИСИМОСТИ (Required)
# ===========================================

# Telegram Bot
python-telegram-bot==21.0.1

# Claude AI API (обновленная версия)
anthropic>=0.74.0

# Utilities
python-dotenv==1.0.1

# Image Processing
Pillow>=10.4.0

# Web Scraping для поиска нормативов
requests==2.31.0
beautifulsoup4==4.12.3

# HTTP клиент для xAI Grok API
httpx==0.27.0

# OpenAI - DALL-E 3 (генерация изображений) и Whisper (распознавание речи)
openai>=1.0.0

# ===========================================
# ОПЦИОНАЛЬНЫЕ ЗАВИСИМОСТИ (Optional)
# ===========================================

# Database - для хранения истории в PostgreSQL
# Без этого используются JSON файлы
asyncpg==0.29.0

# PDF/Word Export - для экспорта истории
# Без этого экспорт в PDF недоступен (только Word)
reportlab==4.0.7
python-docx==1.1.0

# Redis Cache - для кэширования ответов
# Без этого кэш работает только в памяти
redis==5.0.1

# PDF Processing - для обработки PDF документов
PyPDF2==3.0.1

# Vosk - офлайн распознавание речи (fallback для голосовых сообщений)
# Без этого голос работает только через OpenAI Whisper
# Требуется также скачать модель: vosk-model-small-ru-0.22
vosk>=0.3.45

# ===========================================
# ДОПОЛНИТЕЛЬНЫЕ (Extra)
# ===========================================

# Logging (optional)
# python-json-logger==2.0.7
